Adding TypeScript support
