+++ package.json
@@ -18,5 +18,12 @@
   "license": "MIT",
   "devDependencies": {
     "snowpack": "^3.0.11"
+  },
+  "dependencies": {
+    "@types/jest": "^27.4.0",
+    "@types/node": "^17.0.8",
+    "@types/react": "^17.0.38",
+    "@types/react-dom": "^17.0.11",
+    "typescript": "^4.5.4"
   }
 }
